+++ packages/browser/src/browser/__tests__/standalone-analytics.test.ts
@@ -1,13 +1,14 @@
 import jsdom, { JSDOM } from 'jsdom'
-import { InitOptions } from '../../'
+import { InitOptions, getGlobalAnalytics } from '../../'
 import { AnalyticsBrowser, loadLegacySettings } from '../../browser'
 import { snippet } from '../../tester/__fixtures__/segment-snippet'
-import { install, AnalyticsStandalone } from '../standalone-analytics'
+import { install } from '../standalone-analytics'
 import unfetch from 'unfetch'
 import { PersistedPriorityQueue } from '../../lib/priority-queue/persisted'
 import { sleep } from '../../lib/sleep'
 import * as Factory from '../../test-helpers/factories'
 import { EventQueue } from '../../core/queue/event-queue'
+import { AnalyticsStandalone } from '../standalone-interface'
 
 const track = jest.fn()
 const identify = jest.fn()
@@ -142,7 +143,7 @@ describe('standalone bundle', () => {
       .mockImplementation((): Promise<Response> => fetchSettings)
     const mockCdn = 'http://my-overridden-cdn.com'
 
-    window.analytics._cdn = mockCdn
+    getGlobalAnalytics()!._cdn = mockCdn
     await loadLegacySettings(segmentDotCom)
 
     expect(unfetch).toHaveBeenCalledWith(expect.stringContaining(mockCdn))
@@ -262,7 +263,7 @@ describe('standalone bundle', () => {
 
     // register is called after flushPreBuffer in `loadAnalytics`
     register.mockImplementationOnce(() =>
-      window.analytics.track('race conditions', { foo: 'bar' })
+      getGlobalAnalytics()?.track('race conditions', { foo: 'bar' })
     )
 
     await install()