+++ packages/browser/src/browser/standalone-analytics.ts
@@ -1,30 +1,19 @@
-import { Analytics, InitOptions } from '../core/analytics'
 import { AnalyticsBrowser } from '.'
 import { embeddedWriteKey } from '../lib/embedded-write-key'
-
-export interface AnalyticsSnippet extends AnalyticsStandalone {
-  load: (writeKey: string, options?: InitOptions) => void
-}
-
-export interface AnalyticsStandalone extends Analytics {
-  _loadOptions?: InitOptions
-  _writeKey?: string
-  _cdn?: string
-}
-
-declare global {
-  interface Window {
-    analytics: AnalyticsSnippet
-  }
-}
+import { AnalyticsSnippet } from './standalone-interface'
+import {
+  getGlobalAnalytics,
+  setGlobalAnalytics,
+} from '../lib/global-analytics-helper'
 
 function getWriteKey(): string | undefined {
   if (embeddedWriteKey()) {
     return embeddedWriteKey()
   }
 
-  if (window.analytics._writeKey) {
-    return window.analytics._writeKey
+  const analytics = getGlobalAnalytics()
+  if (analytics?._writeKey) {
+    return analytics._writeKey
   }
 
   const regex = /http.*\/analytics\.js\/v1\/([^/]*)(\/platform)?\/analytics.*/
@@ -59,7 +48,7 @@ function getWriteKey(): string | undefined {
 
 export async function install(): Promise<void> {
   const writeKey = getWriteKey()
-  const options = window.analytics?._loadOptions ?? {}
+  const options = getGlobalAnalytics()?._loadOptions ?? {}
   if (!writeKey) {
     console.error(
       'Failed to load Write Key. Make sure to use the latest version of the Segment snippet, which can be found in your source settings.'
@@ -67,8 +56,7 @@ export async function install(): Promise<void> {
     return
   }
 
-  window.analytics = (await AnalyticsBrowser.standalone(
-    writeKey,
-    options
-  )) as AnalyticsSnippet
+  setGlobalAnalytics(
+    (await AnalyticsBrowser.standalone(writeKey, options)) as AnalyticsSnippet
+  )
 }