+++ packages/browser/src/index.ts
@@ -7,5 +7,7 @@ export * from './core/events'
 export * from './core/plugin'
 export * from './core/user'
 
-export type { AnalyticsSnippet } from './browser/standalone-analytics'
+export type { AnalyticsSnippet } from './browser/standalone-interface'
 export type { MiddlewareFunction } from './plugins/middleware'
+export { getGlobalAnalytics } from './lib/global-analytics-helper'
+export { UniversalStorage, Store, StorageObject } from './core/storage'