+++ packages/consent/consent-tools-integration-tests/src/page-bundles/consent-tools-vanilla/index.ts
@@ -1,0 +1,17 @@
+import { AnalyticsBrowser } from '@segment/analytics-next'
+import { createWrapper } from '@segment/analytics-consent-tools'
+
+const fakeCategories = { FooCategory1: true, FooCategory2: true }
+
+const withCMP = createWrapper({
+  getCategories: () => fakeCategories,
+})
+
+const analytics = new AnalyticsBrowser()
+
+withCMP(analytics).load({
+  writeKey: 'foo',
+})
+
+// for testing
+;(window as any).analytics = analytics