+++ packages/consent/consent-tools-integration-tests/src/page-objects/basic.ts
@@ -1,9 +1,0 @@
-import { BasePage } from './base-page'
-
-class BasicPage extends BasePage {
-  constructor() {
-    super('basic.html')
-  }
-}
-
-export default new BasicPage()