+++ packages/consent/consent-tools-integration-tests/src/page-objects/consent-tools-vanilla.ts
@@ -1,0 +1,9 @@
+import { BasePage } from './base-page'
+
+class ConsentToolsVanilla extends BasePage {
+  constructor() {
+    super('consent-tools-vanilla.html')
+  }
+}
+
+export default new ConsentToolsVanilla()