+++ packages/consent/consent-tools/src/domain/validation/validation-error.ts
@@ -1,10 +1,11 @@
 import { AnalyticsConsentError } from '../../types/errors'
 
 export class ValidationError extends AnalyticsConsentError {
-  constructor(message: string, received: any) {
-    super(
-      'ValidationError',
-      `[Validation] ${message} (${`Received: ${JSON.stringify(received)})`}`
-    )
+  constructor(message: string, received?: any) {
+    if (arguments.length === 2) {
+      // to ensure that explicitly passing undefined as second argument still works
+      message += ` (Received: ${JSON.stringify(received)})`
+    }
+    super('ValidationError', `[Validation] ${message}`)
   }
 }