+++ packages/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @segment/analytics-node
 
+## 1.1.1
+
+### Patch Changes
+
+- [#946](https://github.com/segmentio/analytics-next/pull/946) [`edfb8b5`](https://github.com/segmentio/analytics-next/commit/edfb8b5c4463c2ccd336fdfc7c35d4cd711f5410) Thanks [@danieljackins](https://github.com/danieljackins)! - Include sentAt field in payload
+
+- Updated dependencies [[`ee855ba`](https://github.com/segmentio/analytics-next/commit/ee855bad751c393a40dcbde7ae861f27d2b4da26)]:
+  - @segment/analytics-core@1.3.1
+
 ## 1.1.0
 
 ### Minor Changes